Cambios 25/02
-Añadido módulo de seguridad.
-Añadido recoger usuarios de la BDD para acceso a la app.
-Cambio en la manera de registrar a los usuarios mediante ResponseEntity.
-Se ha quitado el video de fondo de index.html para poder hacer el commit.

+++ src/main/java/prw/stwr/repository/GalaxyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import prw.stwr.model.Galaxy;
 
 @Repository
-public interface GalaxyRepository extends JpaRepository<Galaxy, Integer>{
-
+public interface GalaxyRepository extends JpaRepository<Galaxy, Long>{
+	
 }
